refactor/#108: 타입 정의 파일 경로 수정 및 불필요한 API 호출 제거
