+++ TinyBite/app/(auth)/signup/region.tsx
@@ -5,8 +5,8 @@ import { useAuthStore } from "@/stores/authStore";
 import { TermCode, useSignupStore } from "@/stores/signupStore";
 import { colors } from "@/styles/colors";
 import { textStyles } from "@/styles/typography/textStyles";
-import { ApiError } from "@/types/api";
-import { SignupRespone } from "@/types/auth";
+import { ApiError } from "@/types/api.types";
+import { SignupRespone } from "@/types/auth.types";
 import { getErrorMessage } from "@/utils/getErrorMessage";
 import { useMutation } from "@tanstack/react-query";
 import { AxiosError } from "axios";
