+++ TinyBite/app/(auth)/signup/terms.tsx
@@ -6,7 +6,7 @@ import { TermCode, useSignupStore } from "@/stores/signupStore";
 import { useTimerStore } from "@/stores/timerStore";
 import { colors } from "@/styles/colors";
 import { textStyles } from "@/styles/typography/textStyles";
-import { ApiError } from "@/types/api";
+import { ApiError } from "@/types/api.types";
 import { getErrorMessage } from "@/utils/getErrorMessage";
 import { useMutation } from "@tanstack/react-query";
 import { AxiosError } from "axios";
