+++ TinyBite/app/party-detail/[id].tsx
@@ -1,5 +1,4 @@
 import { deleteParty } from "@/api/partyApi";
-import { getUserMe } from "@/api/userApi";
 import ConfirmModal from "@/components/ConfirmModal";
 import PartyDetailBackButton from "@/components/main/party-detail/PartyDetailBackButton";
 import PartyDetailCTA from "@/components/main/party-detail/PartyDetailCTA";
@@ -12,10 +11,11 @@ import PartyDetailPill from "@/components/main/party-detail/PartyDetailPill";
 import PartyDetailProductLink from "@/components/main/party-detail/PartyDetailProductLink";
 import { usePartyDetail } from "@/hooks/usePartyDetail";
 import { useUserCoords } from "@/hooks/useUserCoords";
+import { useAuthStore } from "@/stores/authStore";
 import { useEditPartyStore } from "@/stores/editPartyStore";
 import { colors } from "@/styles/colors";
 import { textStyles } from "@/styles/typography/textStyles";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQueryClient } from "@tanstack/react-query";
 import { useLocalSearchParams, useRouter } from "expo-router";
 import { StatusBar } from "expo-status-bar";
 import React, { useEffect, useState } from "react";
@@ -48,10 +48,7 @@ export default function PartyDetailScreen() {
   );
 
   // 현재 로그인한 사용자 정보 조회
-  const { data: userMe } = useQuery({
-    queryKey: ["getUserMe"],
-    queryFn: getUserMe,
-  });
+  const { user } = useAuthStore();
 
   // 위치 정보 가져오기
   const { coords, refresh: fetchCoords } = useUserCoords();
@@ -139,7 +136,7 @@ export default function PartyDetailScreen() {
         {/* 뒤로가기 버튼 - 헤더 위에 고정 */}
         <PartyDetailBackButton />
         {/* 더보기 버튼 (오른쪽) - 헤더 위에 고정 - 본인이 작성한 파티일 때만 표시 */}
-        {partyDetail?.host.userId === userMe?.userId && (
+        {partyDetail?.host.userId === user?.userId && (
           <PartyDetailMoreButton
             onEdit={() => {
               router.navigate({
