+++ TinyBite/app/party/edit/[type].tsx
@@ -9,8 +9,8 @@ import GlobalButton from "@/components/GlobalButton";
 import { Photo } from "@/stores/creatingPartyStore";
 import { PhotoUrl, useEditPartyStore } from "@/stores/editPartyStore";
 import { colors } from "@/styles/colors";
-import { ApiError } from "@/types/api";
-import { EditedPartyInfo } from "@/types/party";
+import { ApiError } from "@/types/api.types";
+import { EditedPartyInfo } from "@/types/party.types";
 import { getErrorMessage } from "@/utils/getErrorMessage";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { AxiosError } from "axios";
