+++ TinyBite/components/main/MainCard.tsx
@@ -1,6 +1,6 @@
 import { colors } from "@/styles/colors";
 import { textStyles } from "@/styles/typography/textStyles";
-import { PartyItem } from "@/types/party";
+import { PartyItem } from "@/types/party.types";
 import { useState } from "react";
 import {
   Image,
