+++ TinyBite/components/main/MainHeader.tsx
@@ -1,7 +1,6 @@
-import { getUserMe } from "@/api/userApi";
+import { useAuthStore } from "@/stores/authStore";
 import { colors } from "@/styles/colors";
 import { textStyles } from "@/styles/typography/textStyles";
-import { useQuery } from "@tanstack/react-query";
 import { useState } from "react";
 import { Image, StyleSheet, Text, View } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
@@ -18,14 +17,11 @@ interface MainHeaderProps {}
  */
 const MainHeader = ({}: MainHeaderProps = {}) => {
   // 유저 정보 조회
-  const { data: userMe } = useQuery({
-    queryKey: ["getUserMe"],
-    queryFn: getUserMe,
-  });
+  const { user } = useAuthStore();
 
   // 유저 정보에서 닉네임과 위치 가져오기
-  const nickname = userMe?.name || "한입만";
-  const location = userMe?.location || "위치 없음";
+  const nickname = user?.nickname || "한입만";
+  const location = user?.location || "위치 없음";
 
   // 현재 캐러셀 페이지 인덱스 상태 관리
   const [currentPage, setCurrentPage] = useState(0);
